# [canvas] cache cell strings