Added some basic caching for the string representations of cells. Also improved benchmarks to include pipe and text cells. Results:

BEFORE:
```
BenchmarkRender/4x4_no_updates-4                   50000             25880 ns/op
BenchmarkRender/4x4_update_non-transparent-4               30000             58155 ns/op
BenchmarkRender/4x4_update_transparent-4                   30000             57210 ns/op
BenchmarkRender/10x10_update_transparent-4                  5000            347892 ns/op
BenchmarkRender/20x20_update_transparent-4                  1000           1321545 ns/op
BenchmarkRender/50x50_update_transparent-4                   200           8150690 ns/op
BenchmarkRender/100x100_update_transparent-4                  50          32910036 ns/op
BenchmarkRender/4x4_boxed_update_non-transparent-4         20000             94382 ns/op
BenchmarkRender/4x4_boxed_update_transparent-4             20000             93629 ns/op
BenchmarkRender/10x10_boxed_update_transparent-4            3000            416489 ns/op
BenchmarkRender/20x20_boxed_update_transparent-4            1000           1471571 ns/op
BenchmarkRender/50x50_boxed_update_transparent-4             200           8483265 ns/op
BenchmarkRender/100x100_boxed_update_transparent-4            50          33050809 ns/op
```

AFTER:
```
BenchmarkRender/4x4_no_updates-4                   50000             25798 ns/op
BenchmarkRender/4x4_update_non-transparent-4               50000             30297 ns/op
BenchmarkRender/4x4_update_transparent-4                   50000             32276 ns/op
BenchmarkRender/10x10_update_transparent-4                 10000            176013 ns/op
BenchmarkRender/20x20_update_transparent-4                  2000            675142 ns/op
BenchmarkRender/50x50_update_transparent-4                   300           4046671 ns/op
BenchmarkRender/100x100_update_transparent-4                 100          16444711 ns/op
BenchmarkRender/4x4_boxed_update_non-transparent-4         30000             57244 ns/op
BenchmarkRender/4x4_boxed_update_transparent-4             30000             55932 ns/op
BenchmarkRender/10x10_boxed_update_transparent-4            5000            227881 ns/op
BenchmarkRender/20x20_boxed_update_transparent-4            2000            785008 ns/op
BenchmarkRender/50x50_boxed_update_transparent-4             300           4447274 ns/op
BenchmarkRender/100x100_boxed_update_transparent-4           100          17448057 ns/op
```

Overall the `20x20_boxed_update_transparent` case is the most accurate representation of the actual use case, which means the caching resulted in a roughly 85% reduction in cpu time.